feat: allow adding a slice of a file to an archive
Certain applications may want to add slices of files to the archive. They
will be recorded with a given name and their item position (IP) will be
adjusted based on the offset. That is, writing 8192 bytes of a file starting
at offset 4096 will result in an IP of 0 and SZ of 8192. Upon extraction,
the output will be 8192 bytes written to offset 0 of the resulting file, and
the file will have the name given during archive creation.

cargo test passes